MLOps/backend/src/requirements.txt
fastapi
uvicorn
pandas
numpy
mlflow
scikit-learn
imbalanced-learn
python-dotenv
pydantic
python-multipart
matplotlib
seaborn
# Test dependencies
pytest
httpx
